mcq-extractor/pyproject.toml
[project]
name = "mcq-extractor-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pillow>=11.2.1",
    "pymupdf>=1.26.1",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sympy>=1.14.0",
    "tabulate>=0.9.0",
    "uvicorn>=0.34.3",
]
